Enhance convergence step
Add options to  such as tolerance and domain. The user now is able to choose multiple number of level sets that he may want to verify the convergence. Also, the level sets right now are not only defined by decision variables but they can be polynomials or constants (ex. the shape functions p can be an input and even the 0 level set can be compared between iterations)