update convergence.m
The previous version of convergence.m had an error in an operator, it's fixed after this commit. Also the function poly_diff is right now a static function. The function used to perform the integral is ihnerent from the sos toolbox that it's being used, i.e. for the moment this step can only be performed with sosopt toolbox.