Add termination rules (#13)
* Terminate on convergence flag.

* add convergence

add feature to check convergence of a certain level-set, right now it only works for systems with dimension 2

* add convergence

add feature to check convergence of a certain level-set, right now it only works for systems with dimension 2

* Convergence - integral over hypercube

The integral over R^n is implemented, currently for a window of [-1 1] in each dimension. Still to do: change the variable 'info' such that the state machine function is the same for every step; create options so that the user can specify the tolerance in the convergence and the window used for the integration; another approach to verify polynomial similarity based on different norms

* Enhance convergence step

Add options to  such as tolerance and domain. The user now is able to choose multiple number of level sets that he may want to verify the convergence. Also, the level sets right now are not only defined by decision variables but they can be polynomials or constants (ex. the shape functions p can be an input and even the 0 level set can be compared between iterations)

* Simple termination step implemented

With this additional step one can verify if a certain condition is meet over each cycle, in case it's not the algorithm should stop. The condition is between a certain variable from the previous cycle and the current.

* Update +bisos/+iteration/+setp/convergence.m

all requested changes were implemented into this commit

* update +bisos/+iteration/+step/termination

Fix some erros and change some portions of the code according to some suggestions

* update termination.m

Change code according to suggestions. Also a total different layout of the code has been employed, this allows for an easier understanding of the code and debug procedure.

* update README.md with recent developments in termination.m

Some suggestions were employed and also some updates regarding the termination step

* update convergence.m

The previous version of convergence.m had an error in an operator, it's fixed after this commit. Also the function poly_diff is right now a static function. The function used to perform the integral is ihnerent from the sos toolbox that it's being used, i.e. for the moment this step can only be performed with sosopt toolbox.

* Use step-specific info fields to store information about previous solutions.

* Code improvements.

* update termination.m

some changes to turn termination step more robust and faster - decrease number of nested blocks

* update convergence.m

* update termination.m

* update README.m

change the way options are inserted into convergence step

* update demo.m

change how options are inserted into addconvergence

Co-authored-by: Torbjoern Cunis <torbjoern.cunis@gmail.com>